🗃 updates install and upgrade scripts
