+++ Setup/InstallData.php
@@ -6,9 +6,17 @@ namespace Pagarme\Pagarme\Setup;
 use Magento\Framework\Setup\InstallDataInterface;
 use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\ModuleDataSetupInterface;
+use Magento\Customer\Setup\CustomerSetupFactory;
 
 class InstallData implements InstallDataInterface
 {
+
+    private $customerSetupFactory;
+    public function __construct(
+        CustomerSetupFactory $customerSetupFactory
+    ) {
+        $this->customerSetupFactory = $customerSetupFactory;
+    }
 
     /**
      * {@inheritdoc}
@@ -16,10 +24,11 @@ class InstallData implements InstallDataInterface
     public function install(
         ModuleDataSetupInterface $setup,
         ModuleContextInterface $context
-    )
-    {
+    ) {
         $setup->startSetup();
-        $tableName = $setup->getTable('sales_order_status_state');
+        $tableName = $setup->getTable(
+            'sales_order_status_state'
+        );
 
         if ($setup->getConnection()->isTableExists($tableName) == true) {
             $connection = $setup->getConnection();
@@ -28,6 +37,40 @@ class InstallData implements InstallDataInterface
             $connection->update($tableName, $data, $where);
         }
 
+        $this->addCustomerIdPagarme($setup);
         $setup->endSetup();
+    }
+
+    protected function addCustomerIdPagarme($setup)
+    {
+        $setup->startSetup();
+
+        $customerSetup = $this->customerSetupFactory->create(['setup' => $setup]);
+        $attributeCode = 'customer_id_pagarme';
+        $customerSetup->removeAttribute(\Magento\Customer\Model\Customer::ENTITY, $attributeCode);
+        $customerSetup->addAttribute(
+            'customer',
+            $attributeCode,
+            [
+                'label' => 'Customer Id Pagar.me',
+                'type' => 'varchar',
+                'input' => 'text',
+                'required' => false,
+                'visible' => true,
+                'system' => false,
+                'position' => 200,
+                'sort_order' => 200,
+                'user_defined' => false,
+                'default' => '0',
+            ]
+        );
+
+        $eavConfig = $customerSetup->getEavConfig()->getAttribute('customer', $attributeCode);
+        $eavConfig->setData('used_in_forms', ['adminhtml_customer']);
+        $eavConfig->save();
+
+        $setup->endSetup();
+
+        return $setup;
     }
 }
